plusieurs modifs

+++ td2/td2/src/main/kotlin/edu/spring/td2/repositories/OrganizationRepository.kt
@@ -3,5 +3,5 @@ package edu.spring.td2.repositories
 import edu.spring.td2.entities.Organization
 import org.springframework.data.repository.CrudRepository
 
-interface OrganizationRepository:CrudRepository<Organization, Int> {
+interface OrganizationRepository:CrudRepository<Organization, Long> {
 }
